android-record: Reorganize no longer failing tests.
