+++ android-record/src/test/java/com/fasterxml/jackson/module/androidrecord/failing/RecordCreatorsFailingTest.java
@@ -7,8 +7,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.exc.UnrecognizedPropertyException;
 import com.fasterxml.jackson.module.androidrecord.BaseMapTest;
 
-public class RecordCreatorsFailingTest extends BaseMapTest
-{
+public class RecordCreatorsFailingTest extends BaseMapTest {
   static final class RecordWithAltCtor extends RecordTag {
     private final int id;
     private final String name;
@@ -34,28 +33,28 @@ public class RecordCreatorsFailingTest extends BaseMapTest
 
   private final ObjectMapper MAPPER = newJsonMapper();
 
-    /*
-    /**********************************************************************
-    /* Test methods, alternate constructors
-    /**********************************************************************
-     */
+  /*
+  /**********************************************************************
+  /* Test methods, alternate constructors
+  /**********************************************************************
+  */
 
-    // Fails: Implicit canonical constructor still works too
-    public void testDeserializeWithAltCtor() throws Exception {
-        RecordWithAltCtor value = MAPPER.readValue("{\"id\":2812}",
-                RecordWithAltCtor.class);
-        assertEquals(2812, value.id());
-        assertEquals("name2", value.name());
+  // Fails: Implicit canonical constructor still works too
+  public void testDeserializeWithAltCtor() throws Exception {
+    RecordWithAltCtor value = MAPPER.readValue("{\"id\":2812}",
+            RecordWithAltCtor.class);
+    assertEquals(2812, value.id());
+    assertEquals("name2", value.name());
 
-        // "Implicit" canonical constructor can no longer be used when there's explicit constructor
-        try {
-            MAPPER.readValue("{\"id\":2812,\"name\":\"Bob\"}",
-                    RecordWithAltCtor.class);
-            fail("should not pass");
-        } catch (UnrecognizedPropertyException e) {
-            verifyException(e, "Unrecognized");
-            verifyException(e, "\"name\"");
-            verifyException(e, "RecordWithAltCtor");
-        }
+    // "Implicit" canonical constructor can no longer be used when there's explicit constructor
+    try {
+      MAPPER.readValue("{\"id\":2812,\"name\":\"Bob\"}",
+              RecordWithAltCtor.class);
+      fail("should not pass");
+    } catch (UnrecognizedPropertyException e) {
+      verifyException(e, "Unrecognized");
+      verifyException(e, "\"name\"");
+      verifyException(e, "RecordWithAltCtor");
     }
+  }
 }
